Refactor & reorganize everything (#30)
* Refactor "UserService" into "LumaService"

* Refactor Handlers into Handlers::Auth and Handlers::Utils, removed unused imports

* Rename middleware

* Refactor imports from .env and testing
