+++ .env.example
@@ -6,10 +6,9 @@ POSTGRES_USER=
 POSTGRES_PASSWORD=
 POSTGRES_DB=
 POSTGRES_HOST=
-#POSTGRES_HOST=
 POSTGRES_PORT=
 
-# Backend Configuration 
+# Backend Configuration
 BACKEND_PORT=
 BACKEND_LOG_LEVEL=
 BACKEND_DB_CONNECTION=
@@ -17,6 +16,11 @@ DATABASE_URL=
 BACKEND_JWT_SECRET=
 BACKEND_PINECONE_API_KEY=
 BACKEND_PINECONE_INDEX_NAME=
+BACKEND_OPENAI_API_KEY=
+BACKEND_OPENAI_ORG_ID=
+BACKEND_OPENAI_COMPLETION_MODEL=
+BACKEND_OPENAI_CHAT_MODEL=
+BACKEND_OPENAI_EMBEDDING_MODEL=
 
 # Frontend Configuration
 FRONTEND_PORT=
